Complete code refactor. Classes moved to separate .cs files to aid maintainability and readability. Code activated by buttons now has its own method. General clean-up.
