+++ SVGtoGCODE/FittedVector.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SVGtoGCODE
+{
+    // inherit from vector
+    public class FittedVector : Vector
+    {
+        // Should hold the info for the instance of Vector which has been fitted in a specific printing/work space
+    }
+}
