+++ SVGtoGCODE/GCode.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SVGtoGCODE
+{
+    public class GCode
+    {
+        // Is able to convert instance of FittedVector into GCode that complies with the restrictions given in instance of Workspace.
+        // Should be able to generate, save, and export GCode.
+    }
+}
