+++ SVGtoGCODE/SVGtoGCODE.csproj
@@ -63,6 +63,10 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="FittedVector.cs" />
+    <Compile Include="GCode.cs" />
+    <Compile Include="Vector.cs" />
+    <Compile Include="Workspace.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
