+++ SVGtoGCODE/Workspace.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SVGtoGCODE
+{
+    public class Workspace
+    {
+        // Holds the information about the defined workspace, such as the height and width dimensions.
+    }
+}
